+++ pkg/providers/iam/testdata/TestDataBlockedProperties/missing_group_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_iam_blocked_properties" "test" {
+  contract_id    = "ctr_C-123"
+  ui_identity_id = "user123"
+}