+++ pkg/providers/iam/testdata/TestDataBlockedProperties/missing_ui_identity_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_iam_blocked_properties" "test" {
+  group_id    = 1
+  contract_id = "ctr_C-123"
+}